Edited solutions, generator, and validator (adtech-guide-oval-hard)

+++ adtech-guide-oval-hard/solutions/cpp1_tle.cpp
@@ -2,24 +2,24 @@
 using namespace std;
 typedef long long ll;
 
-char solve(int N, int x, int y) {
-    int curx = 0;
-    int cury = 0;
+char solve(ll N, ll x, ll y) {
+    ll curx = 0;
+    ll cury = 0;
     char cur_cell;
     
-    for(int i = 0; i < N; i++) {
+    for(ll i = 0; i < N; i++) {
         curx = 0;
-        for(int j = 0; j < N - i; j++) {
+        for(ll j = 0; j < N - i; j++) {
             if(x == curx && y == cury) return '-';
             curx++;
         }
 
-        for(int j = 0; j < 2 * (N + i); j++) {
+        for(ll j = 0; j < 2 * (N + i); j++) {
             if(x == curx && y == cury) return '#';
             curx++;
         }
 
-        for(int j = 0; j < N - i; j++) {
+        for(ll j = 0; j < N - i; j++) {
             if(x == curx && y == cury) return '-';
             curx++;
         }
@@ -27,19 +27,19 @@ char solve(int N, int x, int y) {
         cury++;
     }
 
-    for(int i = 0; i < N; i++) {
+    for(ll i = 0; i < N; i++) {
         curx = 0;
-        for(int j = 0; j < N; j++) {
+        for(ll j = 0; j < N; j++) {
             if(x == curx && y == cury) return '#';
             curx++;
         }
 
-        for(int j = 0; j < 2 * N; j++) {
+        for(ll j = 0; j < 2 * N; j++) {
             if(x == curx && y == cury) return '-';
             curx++;
         }
 
-        for(int j = 0; j < N; j++) {
+        for(ll j = 0; j < N; j++) {
             if(x == curx && y == cury) return '#';
             curx++;
         }
@@ -47,19 +47,19 @@ char solve(int N, int x, int y) {
         cury++;
     }
 
-    for(int i = 0; i < N; i++) {
+    for(ll i = 0; i < N; i++) {
         curx = 0;
-        for(int j = 0; j < i + 1; j++) {
+        for(ll j = 0; j < i + 1; j++) {
             if(x == curx && y == cury) return '-';
             curx++;
         }
 
-        for(int j = 0; j < 4 * N - 2 * (i + 1); j++) {
+        for(ll j = 0; j < 4 * N - 2 * (i + 1); j++) {
             if(x == curx && y == cury) return '#';
             curx++;
         }
 
-        for(int j = 0; j < i + 1; j++) {
+        for(ll j = 0; j < i + 1; j++) {
             if(x == curx && y == cury) return '-';
             curx++;
         }
